float32 looks to be required.

+++ elephas/java/ndarray.py
@@ -25,7 +25,7 @@ def get_context_dtype():
     Returns the nd4j dtype
     """
 
-    dtype = DataTypeUtil.get_class().getDtypeFromContext('double')
+    dtype = DataTypeUtil.get_class().getDtypeFromContext('float')
     DataTypeUtil.get_class().setDTypeForContext(dtype)
 
     dtype = DataTypeUtil.get_class().getDtypeFromContext()
